+++ packages/react-native-sdk/src/utils/hooks/index.ts
@@ -1,3 +1,2 @@
-export * from './useAppStateListener';
 export * from './useDebouncedValue';
 export * from './usePrevious';