+++ packages/video-filters-react-native/ios/VideoFiltersReactNative-Bridging-Header.h
@@ -1,2 +1,4 @@
 #import <React/RCTBridgeModule.h>
-#import <React/RCTViewManager.h>
+#import <React/RCTUtils.h>
+#import "ProcessorProvider.h"
+