+++ packages/video-filters-web/src/createRenderer.ts
@@ -3,7 +3,7 @@ import { buildWebGL2Pipeline } from './webgl2/webgl2Pipeline';
 import { getSegmentationParams, SegmentationLevel } from './segmentation';
 
 export type BackgroundFilter = 'blur' | 'image';
-export type BackgroundBlurLevel = 'low' | 'medium' | 'high';
+export type BackgroundBlurLevel = 'low' | 'medium' | 'high' | number;
 export type Renderer = {
   /**
    * Disposes of the renderer.