+++ packages/video-filters-react-native/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.2.1](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-filters-react-native-0.2.0...@stream-io/video-filters-react-native-0.2.1) (2024-06-12)
+
+
+### Features
+
+* docs to make video filters usable without the RN SDK ([6061f4d](https://github.com/GetStream/stream-video-js/commit/6061f4d5b83d1ed46051dde12c7d3e269ec26aeb))
+
+## [0.2.0](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-filters-react-native-0.1.1...@stream-io/video-filters-react-native-0.2.0) (2024-06-12)
+
+
+### ⚠ BREAKING CHANGES
+
+* **react-native:** add ios video filters (#1400)
+
+### Features
+
+* **react-native:** add ios video filters ([#1400](https://github.com/GetStream/stream-video-js/issues/1400)) ([dbad806](https://github.com/GetStream/stream-video-js/commit/dbad806e136de7d60a10d292431c8cfe74bd28f9))
+
 ### [0.1.1](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-filters-react-native-0.1.0...@stream-io/video-filters-react-native-0.1.1) (2024-06-06)